requirements-dev.txt
# Install the package itself with all extras
-e ".[headless,socialaccount,mfa,idp-oidc,openid,steam,saml]"

Sphinx
sphinx_rtd_theme

# Testing tools
pytest>=7.4
pytest-asyncio==0.23.8
pytest-django>=4.5.2
pytest-cov

# Code quality
isort==5.13.2
bandit>=1.8.6,<2
black==24.4.0
djlint==1.34.1
flake8>=7.3.0,<8
mypy==1.10.0

# Type stubs
django-stubs==5.0.2
types-requests==2.32.0.20240602
oauthlib>=3.3.0,<4
Pillow>=9.0
psycopg[binary]>=3.2.10
djangorestframework>=3.15.2,<4
django-ninja>=1.3.0,<2
pyyaml>=6.0.2,<7
nox

# IDE
python-lsp-server==1.13.1  # https://pypi.org/project/python-lsp-server/
pylsp-rope==0.1.17  # https://pypi.org/project/pylsp-rope/

# Release
build>=1.3.0,<2
twine>=6.2.0,<7
